Imported TMP + Other Changes
Here is the changes/fixes:

Added ability to drag paint a color across the canvas.
Fixed the screen tearing when panning.
Fixed the "maximum pan" issue (When zoomed in fully to the quad you can't "pan out of it")
Made it such that when zooming, the zoom goes towards where the cursor is, instead of straight down.